requirements.txt
streamlit>=1.24.0
pandas>=1.3.0
numpy>=1.20.0
pillow>=9.0.0
gdown>=4.7.1
matplotlib>=3.5.0
seaborn>=0.12.0
gspread>=5.7.0
google-auth>=2.16.0
pytz
plotly

